refactor: Move cleanEmptyProperties to schema_generation and update references
